requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.2.1
django-cors-headers==2.5.3
django-filter==2.1.0
django-stubs==0.10.2
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==0.4.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
typed-ast==1.3.5
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.3
